apps/note-backend/backend/requirements.txt
aiosqlite==0.19.0
annotated-types==0.7.0
anyio==4.8.0
black==24.1.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
environs==14.1.0
fastapi==0.109.0
fastapi_cors==0.0.6
flake8==7.0.0
Flask>=2.0.1
Flask-CORS>=3.0.10
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-texttospeech>=2.5.2
google-cloud-vision>=2.4.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.24.0
opencv-python==4.11.0.86
openai==1.9.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.26.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.1
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==3.2.0
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv>=0.21.0
python-multipart==0.0.6
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy>=1.4.23
starlette==0.35.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.27.0
Werkzeug==3.1.3
